refactor/test: remove duplicates from tests and clean them
